refactor(ssr): move render functions to own file
